requirements.txt
numpy==1.15.1
scipy==0.19.0
pillow==4.0.0
matplotlib==2.0.0
h5py==2.7.0
cython==0.25.2
pandas==0.19.0
pytest==3.0.7
scikit-image==0.13.0
jupyter==1.0.0
tqdm==4.11.2
munkres==1.0.12
cython==0.25.2
visdom==0.1.8.8
opencv-python==3.4.4.19
imageio==2.4.1
lmdb==0.94
easydict==1.9
prettytable==0.7.2
PyYAML==3.13
